refactor: ♻️ migrate Project to server/messages module
