+++ packages/core/src/promptrunner.ts
@@ -1,6 +1,6 @@
 // Import necessary modules and functions for handling chat sessions, templates, file management, etc.
 import { executeChatSession, tracePromptResult } from "./chat"
-import { Project } from "./ast"
+import { Project } from "./server/messages"
 import { stringToPos } from "./parser"
 import { arrayify, assert, logError, logVerbose, relativePath } from "./util"
 import { runtimeHost } from "./host"
