+++ packages/core/src/systems.ts
@@ -3,10 +3,10 @@
 // It analyzes script options and the JavaScript source code to determine which systems to include or exclude.
 
 import { uniq } from "es-toolkit"
-import { Project } from "./ast"
 import { arrayify } from "./util"
 import { GenerationOptions } from "./generation"
 import { isToolsSupported } from "./tools"
+import { Project } from "./server/messages"
 
 /**
  * Function to resolve and return a list of systems based on the provided script and project.
