+++ packages/vscode/src/state.ts
@@ -9,7 +9,7 @@ import { startLocalAI } from "./localai"
 import { hasOutputOrTraceOpened } from "./markdowndocumentprovider"
 import { pickLanguageModel } from "./lmaccess"
 import { parseAnnotations } from "../../core/src/annotations"
-import { Project } from "../../core/src/ast"
+import { Project } from "../../core/src/server/messages"
 import { JSONLineCache } from "../../core/src/cache"
 import { ChatCompletionsProgressReport } from "../../core/src/chattypes"
 import { fixPromptDefinitions } from "../../core/src/scripts"
@@ -28,7 +28,6 @@ import { MarkdownTrace } from "../../core/src/trace"
 import { logInfo, groupBy } from "../../core/src/util"
 import { CORE_VERSION } from "../../core/src/version"
 import { Fragment, GenerationResult } from "../../core/src/generation"
-import { parametersToVars } from "../../core/src/parameters"
 import { hash, randomHex } from "../../core/src/crypto"
 import { delay } from "es-toolkit"
